requirements.txt
mysql-connector-python==8.1.0
prettytable==3.8.0
python-dotenv==1.0.0
pytest==7.4.3
black==23.11.0
flake8==6.1.0
pylint==3.0.2 